// Prime/src/app/models/tdr.ts
export class TDR {
  VESSEL_NAME: string = '';
  VOYAGE_NO: string = '';
  POL: string = '';
  TERMINAL: string = '';
  ETA: any = '';
  POB_BERTHING: any = '';
  BERTHED: any = '';
  OPERATION_COMMMENCED: any = '';
  POB_SAILING: any = '';
  SAILED: any = '';
  ETD: any = '';
  ETA_NEXTPORT: any = '';
  CREATED_BY: any = '';
  AGENT_CODE: string = '';
}